+++ core/src/main/java/com/jerrysplugins/veinguard/common/CooldownType.java
@@ -1,0 +1,6 @@
+package com.jerrysplugins.veinguard.common;
+
+public enum CooldownType {
+    BLOCK,
+    ALERT
+}